requirements.txt
streamlit
python-dotenv
langgraph
langchain
langchain-groq
PyPDF2
streamlit-mic-recorder
pandas
matplotlib
supabase
gTTS
fpdf2